Rename OtherData to CommonData
Rename test-data/OtherData.js to CommonData.js
to be a bit more descriptive.
