+++ src/components/OSCALProfile.test.js
@@ -5,7 +5,7 @@ import OSCALProfile from "./OSCALProfile";
 import testOSCALMetadata from "./OSCALMetadata.test";
 import testOSCALBackMatter from "./OSCALBackMatter.test";
 import { parentUrlTestData } from "../test-data/Urls";
-import { profileTestData } from "../test-data/OtherData";
+import { profileTestData } from "../test-data/CommonData";
 
 test("OSCALProfile loads", () => {
   render(<OSCALProfileLoader />);
