+++ src/stories/OSCALResponsibleRoles.stories.js
@@ -3,7 +3,7 @@ import OSCALResponsibleRoles from "../components/OSCALResponsibleRoles";
 import {
   exampleParties,
   responsibleRolesTestData,
-} from "../test-data/OtherData";
+} from "../test-data/CommonData";
 
 export default {
   title: "Components/Responsible Roles",
